functions-mcp-server/requirements.txt
azure-functions
requests>=2.31.0
